Update project to bulk copy Vendor directory
The current build fails because the mongo project replaced `mongosniff` with `mongoreplay`. Instead of fixing _the error_ I changed the project to reference `Vendor` as a folder.  I then changed the copy files build stage to copy the _entire_ `Vendor` directory; regardless of contents.

Fixes current build and ensure that future builds work in the face of changes to ancillary distribution package changes.