Change to automatic provisioning / no explicit team
Changing to automatic provisioning makes it easier for contributors to build. With the previous setup people other than the maintainer had to go erase options in the `Build Settings` tab and manage the differences.

With this change, anybody, including the maintainer, just needs to select their team and the projecy builds/signs proproperly. With Xcode 9 works in xcodebuilda and the team can be specified on the command line for CI builds.